Modal e Cadastro com estilização componetizada
